requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
discord.py==1.7.3
requests==2.26.0

